requirements.txt
gin-config
tensorflow-datasets
setproctitle
scipy
meta-model@ git+git://github.com/jackd/meta-model.git@master
pandas
